fix(build): build_server.sh

+++ build_server.sh
@@ -11,17 +11,18 @@ curl -L -O ${url}
 tar xzf ${tag}.tar.gz
 cd ./pyright-${tag}
 
-npm run install:all && npm run build:server && cd ..
+npm run install:all && npm run build:server
 
-test $? -ne 0 && echo "npm build error"
+test $? -ne 0 && echo "npm build error" && exit
 
 echo "Pyright build success!"
 
+cd ..
 rm -rf server/*
 rm -rf typeshed-fallback/*
 
-cp -R ./pyright-${tag}/client/server .
-cp -R ./pyright-${tag}/client/typeshed-fallback .
+cp -R ./pyright-${tag}/client/server/ ./server
+cp -R ./pyright-${tag}/client/typeshed-fallback/ ./typeshed-fallback
 
 rm -rf ${tag}.tar.gz
 rm -rf ./pyright-${tag}
